move tab files
