+++ src/pokemon-details/Modal.tsx
@@ -12,14 +12,14 @@ import { fileTypeFromString } from '../types/FileImport'
 import { PKMInterface } from '../types/interfaces'
 import { OHPKM } from '../types/pkm/OHPKM'
 import FileTypeSelect from './components/FileTypeSelect'
-import NotesDisplay from './NotesTab'
-import OtherDisplay from './OtherTab'
-import RibbonsDisplay from './RibbonsTab'
-import StatsDisplay from './StatsTab'
 import './style.css'
-import SummaryDisplay from './SummaryTab'
 import MetDataMovesTab from './tabs/MetDataMovesTab'
-import TrainersDisplay from './TrainersTab'
+import NotesDisplay from './tabs/NotesTab'
+import OtherDisplay from './tabs/OtherTab'
+import RibbonsDisplay from './tabs/RibbonsTab'
+import StatsDisplay from './tabs/StatsTab'
+import SummaryDisplay from './tabs/SummaryTab'
+import TrainersDisplay from './tabs/TrainersTab'
 
 const PokemonDetailsModal = (props: {
   mon?: PKMInterface
