+++ src/pokemon-details/components/FileTypeSelect.tsx
@@ -34,7 +34,6 @@ const fileTypeColors: Record<string, string> = {
   PA8: '#8cc',
   PK9: '#f52',
   OhpkmV1: 'fuschia',
-  OhpkmV2: 'fuschia',
 }
 
 interface FileTypeSelectProps {
@@ -81,7 +80,6 @@ const FileTypeSelect = (props: FileTypeSelectProps) => {
         <option value={baseFormat}>{baseFormat}</option>
       )}
       <option value="OhpkmV1">OpenHome V1</option>
-      <option value="OhpkmV2">OpenHome V2</option>
     </select>
   )
 }
