+++ src/pokemon-details/tabs/NotesTab.tsx
