+++ src/pokemon-details/tabs/SummaryTab.tsx
@@ -3,17 +3,17 @@ import { getDisplayID } from '@pokemon-files/util'
 import { Badge, Flex, Grid, Spinner, Tooltip } from '@radix-ui/themes'
 import { useMemo } from 'react'
 import { ErrorIcon } from 'src/components/Icons'
-import PokemonIcon from '../components/PokemonIcon'
-import TypeIcon from '../components/TypeIcon'
-import { getPublicImageURL } from '../images/images'
-import { BallsImageList, getItemIconPath } from '../images/items'
-import { PKMInterface } from '../types/interfaces'
-import { getTypes } from '../types/pkm/util'
-import { Styles } from '../types/types'
-import { getTypeColor } from '../util/PokemonSprite'
-import AttributeRow from './components/AttributeRow'
-import AttributeTag from './components/AttributeTag'
-import useMonSprite from './useMonSprite'
+import PokemonIcon from '../../components/PokemonIcon'
+import TypeIcon from '../../components/TypeIcon'
+import { getPublicImageURL } from '../../images/images'
+import { BallsImageList, getItemIconPath } from '../../images/items'
+import { PKMInterface } from '../../types/interfaces'
+import { getTypes } from '../../types/pkm/util'
+import { Styles } from '../../types/types'
+import { getTypeColor } from '../../util/PokemonSprite'
+import AttributeRow from '../components/AttributeRow'
+import AttributeTag from '../components/AttributeTag'
+import useMonSprite from '../useMonSprite'
 
 const styles = {
   column: {
