+++ src/pokemon-details/tabs/TrainersTab.tsx
